Move cleanup method to JiraService
