+++ src/org/ods/orchestration/service/JiraService.groovy
@@ -6,7 +6,6 @@ import com.cloudbees.groovy.cps.NonCPS
 
 import groovy.json.JsonOutput
 import groovy.json.JsonSlurperClassic
-import org.ods.orchestration.util.StringCleanup
 
 import java.net.URI
 
@@ -17,10 +16,23 @@ import org.apache.http.client.utils.URIBuilder
 @SuppressWarnings(['LineLength', 'ParameterName'])
 class JiraService {
 
+    protected static Map CHARACTER_REMOVEABLE = [
+        '\u00A0': ' ',
+    ]
+
     URI baseURL
 
     String username
     String password
+
+    @NonCPS
+    static removeCharacters(inputString) {
+        def outputString = inputString
+        CHARACTER_REMOVEABLE.forEach { k, v ->
+            outputString = outputString.replaceAll(k, v)
+        }
+        return outputString
+    }
 
     JiraService(String baseURL, String username, String password) {
         if (!baseURL?.trim()) {
@@ -261,7 +273,7 @@ class JiraService {
             throw new RuntimeException(message)
         }
 
-        return new JsonSlurperClassic().parseText(StringCleanup.removeCharacters(response.getBody()))
+        return new JsonSlurperClassic().parseText(removeCharacters(response.getBody()))
     }
 
     @NonCPS
@@ -290,7 +302,7 @@ class JiraService {
             throw new RuntimeException(message)
         }
 
-        return new JsonSlurperClassic().parseText(StringCleanup.removeCharacters(response.getBody()))
+        return new JsonSlurperClassic().parseText(removeCharacters(response.getBody()))
     }
 
     @NonCPS
@@ -549,7 +561,7 @@ class JiraService {
             throw new RuntimeException(message)
         }
 
-        return new JsonSlurperClassic().parseText(StringCleanup.removeCharacters(response.getBody()))
+        return new JsonSlurperClassic().parseText(removeCharacters(response.getBody()))
     }
 
     @NonCPS
@@ -719,4 +731,5 @@ class JiraService {
         }
         return true
     }
+
 }
